+++ webapp/src/components/workspaceSwitcher/workspaceOptions.scss
@@ -1,4 +1,4 @@
-.workspaceOption {
+.Option {
     padding: 0px 20px;
     display: flex;
     flex-direction: column;
@@ -28,40 +28,43 @@
     }
 }
 
-.WorkspaceOptions__indicators {
-    color: rgba(var(--center-channel-color-rgb), 0.9);
-    font-size: 18px;
-    border: 2px solid var(--link-color);
-    border-left: none;
-    border-radius: 0 4px 4px 0;
-}
-.WorkspaceOptions__control {
-    border-radius: 8px 8px 0 0 !important;
-    padding: 12px 16px 0;
-    border: none !important;
-}
+.WorkspaceOptions {
+    &__indicators{
+        color: rgba(var(--center-channel-color-rgb), 0.9);
+        font-size: 18px;
+        border: 2px solid var(--link-color);
+        border-left: none;
+        border-radius: 0 4px 4px 0;
+    }
 
-.WorkspaceOptions__control, .WorkspaceOptions__control--is-focused {
-    box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 0px 1px, rgba(0, 0, 0, 0.1) 0px 11px 11px !important;
-}
+    &__control {
+        border-radius: 8px 8px 0 0 !important;
+        padding: 12px 16px 0;
+        border: none !important;
+    }
 
-.WorkspaceOptions__menu {
-    margin-top: 0 !important;
-    border-radius: 0 0 8px 8px !important;
-    border-top: none;
-    box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 0px 1px, rgba(0, 0, 0, 0.1) 0px 11px 11px !important;
-}
+    &__control, &__control--is-focused{
+        box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 0px 1px, rgba(0, 0, 0, 0.1) 0px 11px 11px !important;
+    }
 
-.WorkspaceOptions__value-container {
-    border: 2px solid var(--link-color);
-    border-right: none;
-    border-radius: 4px 0 0 4px;
-}
+    &__menu {
+        margin-top: 0 !important;
+        border-radius: 0 0 8px 8px !important;
+        border-top: none;
+        box-shadow: rgba(0, 0, 0, 0.1) 0px 2px 0px 1px, rgba(0, 0, 0, 0.1) 0px 11px 11px !important;
+    }
 
-.WorkspaceOptions__input, .WorkspaceOptions__placeholder {
-    color: rgb(var(--center-channel-color-rgb)) !important;
-}
+    &__value-container {
+        border: 2px solid var(--link-color);
+        border-right: none;
+        border-radius: 4px 0 0 4px;
+    }
 
-.WorkspaceOptions__control, .WorkspaceOptions__menu {
-    background-color: rgb(var(--center-channel-bg-rgb)) !important;
+    &__input, &__placeholder {
+        color: rgb(var(--center-channel-color-rgb)) !important;
+    }
+
+    &__control, &__menu {
+        background-color: rgb(var(--center-channel-bg-rgb)) !important;
+    }
 }
