+++ webapp/src/octoClient.ts
@@ -383,10 +383,6 @@ class OctoClient {
     }
 
     async getUserWorkspaces(): Promise<UserWorkspace[]> {
-        if (!Utils.isFocalboardPlugin()) {
-            return []
-        }
-
         const path = '/api/v1/workspaces'
         const response = await fetch(this.serverUrl + path, {headers: this.headers()})
         if (response.status !== 200) {
