+++ webapp/src/store/initialLoad.ts
@@ -4,6 +4,14 @@
 import {createAsyncThunk} from '@reduxjs/toolkit'
 
 import {default as client} from '../octoClient'
+import {UserWorkspace} from '../user'
+import {Utils} from '../utils'
+
+const getUserWorkspaces = async ():Promise<UserWorkspace[]> => {
+    // Concept of workspaces is only applicable when running as a plugin.
+    // There is always only one, single workspace in personal server edition.
+    return Utils.isFocalboardPlugin() ? client.getUserWorkspaces() : []
+}
 
 export const initialLoad = createAsyncThunk(
     'initialLoad',
@@ -12,7 +20,7 @@ export const initialLoad = createAsyncThunk(
             client.getWorkspace(),
             client.getWorkspaceUsers(),
             client.getAllBlocks(),
-            client.getUserWorkspaces(),
+            getUserWorkspaces(),
         ])
         if (!workspace) {
             throw new Error('no_workspace')
