+++ webapp/src/store/users.ts
@@ -19,7 +19,6 @@ type UsersStatus = {
     me: IUser|null
     workspaceUsers: {[key: string]: IUser}
     loggedIn: boolean|null
-    userWorkspaces: UserWorkspace[]
 }
 
 const usersSlice = createSlice({
@@ -50,7 +49,6 @@ const usersSlice = createSlice({
                 acc[user.id] = user
                 return acc
             }, {})
-            state.userWorkspaces = action.payload.userWorkspaces
         })
     },
 })
@@ -61,16 +59,10 @@ export const {reducer} = usersSlice
 export const getMe = (state: RootState): IUser|null => state.users.me
 export const getLoggedIn = (state: RootState): boolean|null => state.users.loggedIn
 export const getWorkspaceUsers = (state: RootState): {[key: string]: IUser} => state.users.workspaceUsers
-export const getUserWorkspaces = (state: RootState): UserWorkspace[] => state.users.userWorkspaces
 
 export const getWorkspaceUsersList = createSelector(
     getWorkspaceUsers,
     (workspaceUsers) => Object.values(workspaceUsers).sort((a, b) => a.username.localeCompare(b.username)),
-)
-
-export const getUserWorkspaceList = createSelector(
-    getUserWorkspaces,
-    (userWorkspaces) => userWorkspaces,
 )
 
 export const getUser = (userId: string): (state: RootState) => IUser|undefined => {
